bs4: Specify the markup type for HTML explicitly.
This fixes the following warning when using this library:

```
/srv/zulip-py3-venv/lib/python3.5/site-packages/bs4/__init__.py:185:
UserWarning: No parser was explicitly specified, so I'm using the best
available HTML parser for this system ("lxml"). This usually isn't a
problem, but if you run this code on another system, or in a different
virtual environment, it may use a different parser and behave
differently.

The code that caused this warning is on line 376 of the file
/home/tabbott/zulip/tools/test-backend. To get rid of this warning,
change code that looks like this:

 BeautifulSoup(YOUR_MARKUP)

to this:

 BeautifulSoup(YOUR_MARKUP, "lxml")
```